+++ backend/src/app.ts
@@ -8,7 +8,7 @@ import user from "./routes/user.js";
 import auth from "./routes/auth.js";
 import backgroundFlow from "./routes/backgroundFlow.js";
 import stories from "./routes/stories.js";
-import personas from './routes/personas.js'
+import personas from "./routes/personas.js";
 
 // import stories from './routes/stories.js'
 
@@ -48,7 +48,7 @@ export async function createServer() {
   await fastify.register(characters);
   await fastify.register(backgrounds);
   await fastify.register(stories);
-  // await fastify.register(chats);
+  await fastify.register(chats);
   await fastify.register(user);
   await fastify.register(personas);
   // await fastify.register(tags)
