+++ frontend/src/components/features/card/CardMediaLeft.tsx
