+++ frontend/src/components/features/characterTab/CharacterInfoSection.tsx
