+++ frontend/src/pages/delete/ComponentDemo.tsx
@@ -3,10 +3,10 @@ import { Link } from "react-router-dom";
 import Header from "@/components/layout/Header";
 import FloatingButton from "@/components/features/FloatingButton";
 import BottomNav from "@/components/layout/BottomNav";
-import OtherMessage from "@/components/features/OtherMessage";
-import MyMessage from "@/components/features/MyMessage";
+import OtherMessage from "@/components/features/message/OtherMessage";
+import MyMessage from "@/components/features/message/MyMessage";
 import IconButton from "@/components/ui/IconButton";
-import MessageActions from "@/components/features/MessageActions";
+import MessageActions from "@/components/features/message/MessageActions";
 import { ReactComponent as HeartIcon } from "@/assets/icons/Heart.svg";
 import { ReactComponent as PlayIcon } from "@/assets/icons/Play.svg";
 import { ReactComponent as PlusIcon } from "@/assets/icons/Plus.svg";
